assignment-5/requirements.txt
# Core LangChain dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# Vector database
chromadb==0.4.20

# Environment variables
python-dotenv==1.0.0

# Text processing
tiktoken==0.5.2

# Additional utilities
numpy==1.24.3
pandas==2.0.3 